Add files via upload
Save a copy. Remove the 3rd column (that's notes). Save copied spreadsheet as CSV to make an array.
The first item in the array (col A) is the text to find in the HTML that I'll give you, converted from a PDF.
The second item (col B) is the text to replace it with, or an action to take.